backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
httpx==0.25.1
pytrends==4.9.2
pydantic==2.4.5
urllib3<2.0
